+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/context/trace/TraceSegment.java
@@ -18,11 +18,11 @@
 
 package org.apache.skywalking.apm.agent.core.context.trace;
 
+import com.google.common.collect.Lists;
 import java.util.LinkedList;
 import java.util.List;
 import org.apache.skywalking.apm.agent.core.conf.Config;
 import org.apache.skywalking.apm.agent.core.context.ids.DistributedTraceId;
-import org.apache.skywalking.apm.agent.core.context.ids.DistributedTraceIds;
 import org.apache.skywalking.apm.agent.core.context.ids.GlobalIdGenerator;
 import org.apache.skywalking.apm.agent.core.context.ids.NewDistributedTraceId;
 import org.apache.skywalking.apm.network.language.agent.v3.SegmentObject;
@@ -39,29 +39,14 @@ public class TraceSegment {
     private String traceSegmentId;
 
     /**
-     * The refs of parent trace segments, except the primary one. For most RPC call, {@link #refs} contains only one
-     * element, but if this segment is a start span of batch process, the segment faces multi parents, at this moment,
-     * we use this {@code #refs} to link them.
-     * <p>
-     * This field will not be serialized. Keeping this field is only for quick accessing.
-     */
-    private List<TraceSegmentRef> refs;
-
-    /**
      * The spans belong to this trace segment. They all have finished. All active spans are hold and controlled by
      * "skywalking-api" module.
      */
     private List<AbstractTracingSpan> spans;
 
-    /**
-     * The <code>relatedGlobalTraces</code> represent a set of all related trace. Most time it contains only one
-     * element, because only one parent {@link TraceSegment} exists, but, in batch scenario, the num becomes greater
-     * than 1, also meaning multi-parents {@link TraceSegment}. <p> The difference between
-     * <code>relatedGlobalTraces</code> and {@link #refs} is: {@link #refs} targets this {@link TraceSegment}'s direct
-     * parent, <p> and <p> <code>relatedGlobalTraces</code> targets this {@link TraceSegment}'s related call chain, a
-     * call chain contains multi {@link TraceSegment}s, only using {@link #refs} is not enough for analysis and ui.
-     */
-    private DistributedTraceIds relatedGlobalTraces;
+    private TraceSegmentRef traceSegmentRef;
+
+    private DistributedTraceId relatedGlobalTraceId;
 
     private boolean ignore = false;
 
@@ -75,8 +60,7 @@ public class TraceSegment {
     public TraceSegment() {
         this.traceSegmentId = GlobalIdGenerator.generate();
         this.spans = new LinkedList<>();
-        this.relatedGlobalTraces = new DistributedTraceIds();
-        this.relatedGlobalTraces.append(new NewDistributedTraceId());
+        this.relatedGlobalTraceId = new NewDistributedTraceId();
         this.createTime = System.currentTimeMillis();
     }
 
@@ -86,26 +70,14 @@ public class TraceSegment {
      * @param refSegment {@link TraceSegmentRef}
      */
     public void ref(TraceSegmentRef refSegment) {
-        if (refs == null) {
-            refs = new LinkedList<>();
-        }
-        /*
-         * There maybe an exception in some plugins, which causes the span to fail to exit,
-         * which results in the ThreadLocal not being cleaned up, and this data is increasing.
-         */
-        if (refs.size() >= Config.Agent.TRACE_SEGMENT_REF_LIMIT_PER_SEGMENT) {
-            return;
-        }
-        if (!refs.contains(refSegment)) {
-            refs.add(refSegment);
-        }
+        this.traceSegmentRef = refSegment;
     }
 
     /**
-     * Establish the line between this segment and all relative global trace ids.
+     * Establish the line between this segment and the relative global trace id.
      */
-    public void relatedGlobalTraces(DistributedTraceId distributedTraceId) {
-        relatedGlobalTraces.append(distributedTraceId);
+    public void relatedGlobalTrace(DistributedTraceId distributedTraceId) {
+        this.relatedGlobalTraceId = distributedTraceId;
     }
 
     /**
@@ -128,16 +100,15 @@ public class TraceSegment {
         return traceSegmentId;
     }
 
-    public boolean hasRef() {
-        return !(refs == null || refs.size() == 0);
-    }
-
+    /**
+     * For testing use only.
+     */
     public List<TraceSegmentRef> getRefs() {
-        return refs;
+        return Lists.newArrayList(traceSegmentRef);
     }
 
-    public List<DistributedTraceId> getRelatedGlobalTraces() {
-        return relatedGlobalTraces.getRelatedGlobalTraces();
+    public DistributedTraceId getRelatedGlobalTrace() {
+        return relatedGlobalTraceId;
     }
 
     public boolean isSingleSpanSegment() {
@@ -159,7 +130,7 @@ public class TraceSegment {
      */
     public SegmentObject transform() {
         SegmentObject.Builder traceSegmentBuilder = SegmentObject.newBuilder();
-        traceSegmentBuilder.setTraceId(getRelatedGlobalTraces().get(0).getId());
+        traceSegmentBuilder.setTraceId(getRelatedGlobalTrace().getId());
         /*
          * Trace Segment
          */
@@ -179,7 +150,7 @@ public class TraceSegment {
 
     @Override
     public String toString() {
-        return "TraceSegment{" + "traceSegmentId='" + traceSegmentId + '\'' + ", refs=" + refs + ", spans=" + spans + ", relatedGlobalTraces=" + relatedGlobalTraces + '}';
+        return "TraceSegment{" + "traceSegmentId='" + traceSegmentId + '\'' + ", traceSegmentRef=" + traceSegmentRef + ", spans=" + spans + ", relatedGlobalTraceId=" + relatedGlobalTraceId + '}';
     }
 
     public long createTime() {